get latest copy from split.io
