+++ package.json
@@ -1,7 +1,7 @@
 {
   "name": "split-evaluator",
-  "version": "1.3.1",
-  "description": "Split SDK web services",
+  "version": "1.4.4",
+  "description": "Split-Evaluator microservice",
   "main": "server.js",
   "scripts": {
     "start": "node server.js"
@@ -16,7 +16,7 @@
   },
   "homepage": "https://github.com/splitio/split-evaluator#readme",
   "dependencies": {
-    "@splitsoftware/splitio": "9.3.6",
+    "@splitsoftware/splitio": "10.3.1",
     "config": "1.25.1",
     "express": "4.13.4",
     "lodash": "4.12.0"
